# headlands/src/main/webapp/swagger.json
{"swagger":"2.0","info":{"version":"0.0.4-SNAPSHOT","title":"headlands"},"host":"localhost:8080/headlands","basePath":"/resources","schemes":["http"],"paths":{"/cache-processors/{cache}":{"post":{"consumes":[],"produces":[],"parameters":[{"name":"cache","in":"path","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{}}],"responses":{"204":{"description":"No Content","headers":{}}}}},"/caches":{"get":{"consumes":[],"produces":[],"parameters":[],"responses":{"200":{"description":"OK","headers":{},"schema":{"type":"array","items":{"$ref":"#/definitions/NestedType"}}}}}},"/caches/{cacheName}":{"put":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{"$ref":"#/definitions/CacheConfiguration"}}],"responses":{"200":{"description":"OK","headers":{"x-info":{"type":"string"}}},"201":{"description":"Created","headers":{"Location":{"type":"string"}}}}},"options":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","headers":{},"schema":{"$ref":"#/definitions/CacheConfiguration"}},"204":{"description":"No Content","headers":{}}}}},"/caches/{cacheName}/entries":{"get":{"consumes":[],"produces":["application/json"],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"maxEntries","in":"query","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","headers":{}}}},"put":{"consumes":["application/json"],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{}}],"responses":{"200":{"description":"OK","headers":{"x-message":{"type":"string"}}}}},"delete":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","headers":{}}}}},"/caches/{cacheName}/entries/{key}":{"get":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","headers":{},"schema":{"type":"string"}}}},"put":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","headers":{}}}},"delete":{"consumes":[],"produces":[],"parameters":[{"name":"cacheName","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","headers":{}}}}},"/entry-processors/{cache}":{"post":{"consumes":[],"produces":[],"parameters":[{"name":"cache","in":"path","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{}}],"responses":{"200":{"description":"OK","headers":{},"schema":{"$ref":"#/definitions/JsonObject"}},"204":{"description":"No Content","headers":{"details":{"type":"string"}}},"400":{"description":"Bad Request","headers":{"details":{"type":"string"},"error":{"type":"string"}}},"500":{"description":"Internal Server Error","headers":{"details":{"type":"string"},"error":{"type":"string"}}}}}}},"definitions":{"JsonObject":{"properties":{}},"NestedType":{"properties":{"entries":{"type":"string"},"link":{"type":"string"},"name":{"type":"string"}}},"CacheConfiguration":{"properties":{"storeByValue":{"type":"boolean"},"managementEnabled":{"type":"boolean"},"statisticsEnabled":{"type":"boolean"},"readThrough":{"type":"boolean"},"writeThrough":{"type":"boolean"}}}}}